+++ apps/vetting/apps.py
@@ -1,0 +1,10 @@
+from django.apps import AppConfig
+
+
+class VettingConfig(AppConfig):
+    """*app config for the human vetting workflow*"""
+
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "apps.vetting"
+    label = "vetting"
+    verbose_name = "Vetting"